move upgrate to guest form to settings and cleanup homepage
