+++ app/features/settings/settings.tsx
@@ -5,10 +5,14 @@ import {
   PageFooter,
   PageHeader,
 } from '~/components/page';
+import { Button } from '~/components/ui/button';
 import { Separator } from '~/components/ui/separator';
 import { Money } from '~/lib/money';
 import { canShare, shareContent } from '~/lib/share';
 import { LinkWithViewTransition } from '~/lib/transitions';
+import { UpgradeGuestForm } from '../signup/upgrade-guest-form';
+import { useAuthActions } from '../user/auth';
+import { useUserStore } from '../user/user-provider';
 import { SettingsNavButton } from './components/settings-nav-button';
 
 const username = 'satoshi';
@@ -17,6 +21,9 @@ const activeMintBalance = 10_000;
 const activeMintUnit = 'USD';
 
 export default function Settings() {
+  const { signOut } = useAuthActions();
+  const user = useUserStore((s) => s.user);
+
   const handleShare = async () => {
     const data = {
       title: `${username}'s Boardwalk`,
@@ -67,6 +74,8 @@ export default function Settings() {
         <SettingsNavButton to="/settings/advanced" variant="destructive">
           Advanced
         </SettingsNavButton>
+        <Button onClick={signOut}>Sign Out</Button>
+        {user.isGuest && <UpgradeGuestForm />}
       </PageContent>
 
       <PageFooter>
